Testfile (#36)
* commit for saving

* testutil 작성 + network 나누기
