+++ cmd/network/check_vpc_cni_prefix_mode_test.go
@@ -1,0 +1,1 @@
+package network_test
